+ rename package; + create methods for duplicate checks;
