--- src/main/java/a/gleb/bus_station/Controllers/ClientController.java
@@ -1,12 +1,0 @@
-package a.gleb.bus_station.Controllers;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class ClientController {
-
-
-
-
-
-}
